Convert testsuite to use flask app fixtures
Instead of using a test class which provides helpers via a "self"
object, use the flask app fixtures.

Additionally:
- move the unit tests into the unit testing tree for now
- deduplicate test module names
- remove empty __init__.py files (tests are not necessarily a package)
- remove excess imports throughout one of the test modules

This does not make any changes to test logic, it just moves the test
to the new testing infrastructure.

Because tests do not repeatedly reconstruct the flask application, the
testsuite runs approximately five times faster in this way.
